Sync Bus setting

+++ Remembrance.Core/Sync/SharedRepositoryCloner.cs
@@ -2,7 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Common.Logging;
 using JetBrains.Annotations;
+using Remembrance.Contracts.DAL.Local;
 using Remembrance.Contracts.DAL.Shared;
 using Remembrance.Contracts.Sync;
 using Remembrance.Resources;
@@ -14,9 +16,15 @@ namespace Remembrance.Core.Sync
     internal sealed class SharedRepositoryCloner : ISharedRepositoryCloner, IDisposable
     {
         private readonly IDictionary<ISharedRepository, IRateLimiter> _cloneableRepositoriesWithRateLimiters;
+        private readonly ILocalSettingsRepository _localSettingsRepository;
+        private readonly ILog _logger;
 
-        public SharedRepositoryCloner([NotNull] IReadOnlyCollection<ISharedRepository> cloneableRepositories, [NotNull] Func<IRateLimiter> rateLimiterFactory)
+        public SharedRepositoryCloner([NotNull] IReadOnlyCollection<ISharedRepository> cloneableRepositories, [NotNull] Func<IRateLimiter> rateLimiterFactory,
+            [NotNull] ILocalSettingsRepository localSettingsRepository,
+            [NotNull] ILog logger)
         {
+            _localSettingsRepository = localSettingsRepository ?? throw new ArgumentNullException(nameof(localSettingsRepository));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _ = cloneableRepositories ?? throw new ArgumentNullException(nameof(cloneableRepositories));
             _ = rateLimiterFactory ?? throw new ArgumentNullException(nameof(rateLimiterFactory));
             _cloneableRepositoriesWithRateLimiters = cloneableRepositories.ToDictionary(cloneableRepository => cloneableRepository, cloneableRepository => rateLimiterFactory());
@@ -40,19 +48,33 @@ namespace Remembrance.Core.Sync
             var repository = (ISharedRepository)sender;
             var rateLimiter = _cloneableRepositoriesWithRateLimiters[repository];
             rateLimiter.Throttle(
-                TimeSpan.FromMinutes(1),
+                TimeSpan.FromSeconds(5),
                 () =>
                 {
-                    var fileName = repository.DbFileName + repository.DbFileExtension;
-                    var oldFilePath = Path.Combine(repository.DbDirectoryPath, fileName);
-                    var newDirectoryPath = RemembrancePaths.SharedDataPath;
-                    var newFilePath = Path.Combine(newDirectoryPath, fileName);
-                    if (File.Exists(newFilePath))
+                    var syncBus = _localSettingsRepository.SyncBus;
+                    if (syncBus == SyncBus.NoSync)
                     {
-                        File.Delete(newFilePath);
+                        return;
                     }
 
-                    File.Copy(oldFilePath, newFilePath);
+                    var fileName = repository.DbFileName + repository.DbFileExtension;
+                    var oldFilePath = Path.Combine(repository.DbDirectoryPath, fileName);
+                    var newFilePath = Path.Combine(RemembrancePaths.GetSharedPath(syncBus), fileName);
+                    try
+                    {
+                        if (File.Exists(newFilePath))
+                        {
+                            File.Delete(newFilePath);
+                        }
+
+                        File.Copy(oldFilePath, newFilePath);
+                        _logger.InfoFormat("Cloned repository {0} to {1}", oldFilePath, newFilePath);
+                    }
+                    catch
+                    {
+                        _logger.WarnFormat("Cannot clone repository {0} to {1}. Retrying...", oldFilePath, newFilePath);
+                        Repository_Changed(sender, e);
+                    }
                 });
         }
     }
